+++ packages/abus/features/Messaging/Cancellation/Cancel_exiting_handler.Spec.ts
@@ -188,8 +188,8 @@ feature(`Cancel existing handler
                 """
                 `, async () => {
                     messageToSend = stepContext.docStringAsEntity;
-                    await waitUntilAsync(() => outboundLogger.messages.length >= messageToSend.length, 200);
-                    await sleep(10); // provide a little buffer to ensure additional messages don't arrive unexpectedly
+                    await waitUntilAsync(() => outboundLogger.messages.length >= messageToSend.length, 500);
+                    await sleep(100); // provide a little buffer to ensure additional messages don't arrive unexpectedly
 
                     const messages = outboundLogger.messages;
                     messages.length.should.eq(messageToSend.length);