+++ packages/abus/docs/release-notes.md
@@ -3,3 +3,4 @@
 
 ### additional with no issue logged
 * changed npm package to abus from abus2
+* added method `waitForEventAsync` to ISendMessages interface. This allows the ability to create a temporary handler for an event.